Extract whole game test logic to library.
